paging library 3 testing

+++ app/src/main/java/com/example/testingproject/models/FavNewsModel.kt
@@ -14,7 +14,7 @@ class FavNewsModel(
 {
     @PrimaryKey(autoGenerate = true)
     @NotNull
-    var favnewsId: Int? = null
+    var favnewsId: Long? = null
 
 
 }
